+++ Scripts/CommonServerPython/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## [Unreleased]
+Added the ***parse_date_string*** function, which parses the date string to a datetime object.
+
+## [19.9.0] - 2019-09-04
+  - IntegrationLogger improvements.
   - Added support for IPv6 in the ***is_ip_valid*** command.
   - Added function ***get_demisto_version*** which return the Demisto server version and build number.